requirements.txt
Django==2.2.6
gunicorn==19.9.0
psycopg2-binary==2.8.3
django_rest_framework
django-import-export
django-cors-headers
whitenoise==3.3.1